+++ src/components/sections/SMHowItWorks.css
